protected auth routes

+++ src/controllers/notes.ts
@@ -2,10 +2,16 @@ import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    authenticatedUserId;
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -13,8 +19,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 };
 
 export const getNote: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note id');
     }
@@ -23,6 +31,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
     res.status(200).json(note);
   } catch (error) {
@@ -43,12 +55,15 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, 'Note must have a title');
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -77,8 +92,10 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note Id');
     }
@@ -90,6 +107,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
 
     note.title = newTitle;
@@ -104,8 +125,10 @@ export const updateNote: RequestHandler<
 
 export const deleteNode: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, 'Invalid note id');
     }
@@ -114,6 +137,10 @@ export const deleteNode: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note');
     }
 
     await note.deleteOne();
